Update RC tagging and test process flow to ensure RC testing is proper.
RC testing process which we have now builds and RC image based on last released openshift version
Ex. 4.17.1.rc will be actually testing 4.16.0 helm image.

This flow is not correct.
We missed to see a production issue in rc testing due to this.

So, we are now certifying an rc build first, then update the image in this repo before proceeding with local testing.

This ensures rc build which we test is similar to production build and helps us capture in rc testing.